[impl] generated client and initial views

+++ ComicBooks.Angular/src/app/app.component.ts
@@ -1,11 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { ApiClientService, FloorPlanDto } from '../services/api-client-service';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet],
+  imports: [RouterOutlet, FormsModule],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
 })
@@ -14,16 +15,18 @@ export class AppComponent implements OnInit {
 
   floorPlanName: string = '';
   createdFloorPlan?: FloorPlanDto;
-  floorPlans?: FloorPlanDto[];
+  floorPlans?: FloorPlanDto[] = undefined;
 
-  constructor(private apiClient: ApiClientService) { }
+  constructor(private apiClient: ApiClientService) {
+
+  }
 
   ngOnInit(): void {
     this.loadFloorPlans();
   }
 
   loadFloorPlans(): void {
-    this.apiClient.floorplansAll().subscribe(floorPlans => {
+    this.apiClient.getAllFloorPlans().subscribe(floorPlans => {
       this.floorPlans = floorPlans;
     });
   }
@@ -37,13 +40,24 @@ export class AppComponent implements OnInit {
     });
   }
 
-  deleteFloorPlan(id: string): void {
+  deleteFloorPlan(id: string|undefined): void {
+    if (!id) return;
     this.apiClient.deleteFloorPlan(id).subscribe(() => {
       this.loadFloorPlans();
+    }, (error) => {
+      if (error.status === 204) return;
+      console.error('Error deleting floor plan:', error);
     });
   }
 
-  editFloorPlan(id: string): void {
+  editFloorPlan(plan: FloorPlanDto): void {
+    if (!plan.id) return;
+    this.apiClient.updateFloorPlan(plan.id, plan).subscribe(() => {
+      console.log(`Edit FloorPlan with ID: ${plan.id}`);
+    },
+    (error) => {
+      console.error('Error updating floor plan:', error);
+    });
     // Implement navigation to edit page if needed
     // For example: this.router.navigate(['/floorplans/edit', id]);
   }
